requirements.txt
Django==4.2.5
autoflake==2.2.1
black==23.9.1
flake8==6.1.0
isort==5.12.0
pre-commit==3.4.0
python-dotenv==1.0.0
django-filter==23.3
djoser==2.2.0
drf-extra-fields==3.7.0
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
psycopg2==2.9.7
django-cors-headers==4.2.0
drf-spectacular==0.26.5